samba-build: fix deps for package samba-test
We are building talloc/tdb/tevent/ldb internally.
The actual libs are not shipped in official locations, but the
python3 binding libs are. So add the appropriate provides to
the samba-client-libs package.

This fixes depedency problems when installing the samba-test
package, as this depends on these python3-{talloc,tevent,ldb,tdb}
packages.

Signed-off-by: Michael Adam <obnox@samba.org>