Split out running tests from setup of client machines
At the moment we run the tests in the same playbook as setting up of the
client machines. With this patch we separate out these two distinct
processes.

This will also allow easier development of tests as it reduces the time
required to run the tests on an existing environment.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>