Add separate sections for pip2 and pip3
CentOS 7 has python versions 2.7 and 3.8 installed. When using pip to
install these modules, the module location depends on the version we are
installing for. We use the default python location for all our
python scripts used in the test module but the newer selftest modules
requires python 3. This is making it difficult to maintain using the
right version of modules without running OS specific calls.

To avoid these versioning issues, we plan on moving our scripts to
python 3 and while we transition, we temporarily use pip for both python
2 and python 3. We intend on removing the section for python 2 in the
  near future.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>