Change testscripts to  use Python3
The new selftest helpers require python3. While out own scripts use the
default python interpreter which for CentOS 7 is python2. On systems
with both versions installed, we need to decide which versions to
install a particluar module for.

To avoid confusion due to differring versions, we have decided to
stick to Python3 for all our python scripts.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>