Make sure num_public is int
Since num_public is calculated by dividing the combo number with the
number of test users, the result is a float. num_public is then used as
an index to an array and is expected to be used as an integer.

This while it works with python2, fails with python3. Fix the code to
ensure num_public is an integer.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>