samba-build: fix test-install of centos8 rpms
the talloc and ldb manpages installed witht the client-libs package
conflict with the same files from libtalloc-devel and libldb-devel
respectively. For centos8, this lets install fail since the
samba3-python package indirectly requires these devel packages.

Since we're building the libs as bundled, we should not
be installing these manpages in a system location anyway.

Fix this by not packaging these manpages in the client-libs package.

Signed-off-by: Michael Adam <obnox@samba.org>