Fix inclusion of PyDSDB library files
Commit aeded2d88dc21e5eb0d799181b2d009504f135fd removed the following
PyDSDB library files from buildroot:

* /usr/lib64/python3.9/site-packages/samba/dsdb.cpython-39-x86_64-linux-gnu.so
* /usr/lib64/python3.9/site-packages/samba/dsdb_dns.cpython-39-x86_64-linux-gnu.so

This was done under the impression that their dependency on
libdnsserver-common-samba4.so was not built in a non AD DC
environment. But in reality it was also conditionally removed
from the buildroot.

Therefore we avoid removing libdnsserver-common-samba4.so from buildroot
and subsequently include it in samba-libs to satisfy all dependencies.
Additionally we remove PyDSDB listing from %files section of python3-samba-dc
sub-package.